+++ deploy-forge/deploy-forge.php
@@ -3,7 +3,7 @@
  * Plugin Name: Deploy Forge
  * Plugin URI: https://getdeployforge.com
  * Description: Automates theme deployment from GitHub repositories via Deploy Forge platform
- * Version: 1.0.54
+ * Version: 1.0.55
  * Author: Deploy Forge
  * Author URI: https://getdeployforge.com
  * License: GPL v2 or later